update: upgrade CardWIdgetVerify with size responsive
